feat(compiler): implement compiler identification and parsing
- Add `identify.ts` for compiler identification logic, including custom rule registration and matching.
- Introduce `index.ts` to export types and functions related to compiler operations.
- Create `clang.ts` parser for handling Clang compiler commands with detailed option parsing.
- Implement `gnu.ts` and `msvc.ts` parsers that leverage Clang parsing logic for GNU and MSVC commands.
- Add `nvcc.ts` as a placeholder for NVCC command parsing.
- Define types in `types.ts` for compiler phases, artifacts, and command models.
- Implement `unwrap.ts` to handle command unwrapping.
- Enhance tests in `compiler-cmd.ts` to validate new compiler identification and parsing features, including custom rule registration and replacement.
