+++ api/src/cmd/compiler/index.ts
@@ -1,0 +1,4 @@
+export * from "./types.js";
+export * from "./unwrap.js";
+export * from "./identify.js";
+export * from "./analysis.js";
