+++ api/src/cmd/compiler/parsers/gnu.ts
@@ -1,0 +1,1 @@
+export { parseGnuCommand } from "./clang.js";
