+++ api/src/cmd/compiler/parsers/msvc.ts
@@ -1,0 +1,1 @@
+export { parseMsvcCommand } from "./clang.js";
