+++ api/src/cmd/compiler/parsers/nvcc.ts
@@ -1,0 +1,8 @@
+import type { CommandModel, CompilerIdentity } from "../types.js";
+
+export function parseNvccCommand(
+  _cmd: readonly string[],
+  _identity: CompilerIdentity,
+): CommandModel | undefined {
+  return undefined;
+}
