+++ api/src/cmd/compiler/unwrap.ts
@@ -1,0 +1,10 @@
+import type { UnwrappedCompilerCommand } from "./types.js";
+
+export function unwrapCompilerCommand(
+  argv: readonly string[],
+): UnwrappedCompilerCommand {
+  return {
+    argv: [...argv],
+    originalArgv: [...argv],
+  };
+}
